refactor(mcp): extract MCPDispatcher and add resources/list and resources/read
